fixed popup bugs and completed UTs

+++ src/core/CardTemplate.gd
@@ -25,7 +25,8 @@ enum{
 	FOCUSED_ON_BOARD		#8
 	DROPPING_INTO_PILE 		#9
 	IN_PILE					#10
-	FOCUSED_IN_POPUP		#11
+	IN_POPUP				#11
+	FOCUSED_IN_POPUP		#12
 }
 
 # The possible return codes a function can return
@@ -179,14 +180,13 @@ func _on_Card_mouse_entered() -> void:
 					state = FOCUSED_ON_BOARD
 			# The only way to mouse over a card in a pile, is when
 			# it's in a the grid popup
-			IN_PILE:
+			IN_POPUP:
 					state = FOCUSED_IN_POPUP
 
 
 # A signal for whenever the player clicks on a card
 func _on_Card_gui_input(event) -> void:
 	if event is InputEventMouseButton:
-		print("click")
 		# If the player presses the left click, it might be because
 		# they want to drag the card
 		if event.is_pressed() and event.get_button_index() == 1:
@@ -263,7 +263,7 @@ func _on_Card_mouse_exited() -> void:
 			FOCUSED_ON_BOARD:
 				state = ON_PLAY_BOARD
 			FOCUSED_IN_POPUP:
-				state = IN_PILE
+				state = IN_POPUP
 
 
 # Resizes Token Drawer to min size whenever a token is removed completely.
@@ -425,7 +425,7 @@ func get_targetcard() -> Card:
 # Returns _ReturnCode.CHANGED if the card actually changed rotation
 #
 # Returns _ReturnCode.OK if the card was already in the correct rotation
-func set_is_faceup(value: bool) -> int:
+func set_is_faceup(value: bool, instant := false) -> int:
 	var retcode: int
 	if value == is_faceup:
 		retcode = _ReturnCode.OK
@@ -438,7 +438,7 @@ func set_is_faceup(value: bool) -> int:
 		if set_is_viewed(false) == _ReturnCode.FAILED:
 			print("ERROR: Something went unexpectedly in set_is_faceup")
 		if value:
-			_flip_card($Control/Back, $Control/Front)
+			_flip_card($Control/Back, $Control/Front,instant)
 			$Control/ManipulationButtons/View.visible = false
 			_stop_pulse()
 			# When we flip face up, we also want to show the dupe card
@@ -456,7 +456,7 @@ func set_is_faceup(value: bool) -> int:
 					var dupe_back = dupe_card.get_node("Control/Back")
 					_flip_card(dupe_back, dupe_front, true)
 		else:
-			_flip_card($Control/Front, $Control/Back)
+			_flip_card($Control/Front, $Control/Back,instant)
 			$Control/ManipulationButtons/View.visible = true
 #			if get_parent() == cfc.NMAP.board:
 			_start_pulse()
@@ -625,26 +625,34 @@ func move_to(targetHost: Node2D,
 			if set_is_faceup(true) == _ReturnCode.FAILED:
 				print("ERROR: Something went unexpectedly in set_is_faceup")
 		elif targetHost in cfc.piles:
-			# Added because sometimes it ended up stuck and a card remained
-			# visible on top of deck
-			$Tween.remove_all()
-			# We need to adjust the end position based on the local rect inside
-			# the container control node
-			# So we transform global coordinates to container rect coordinates.
-			previous_pos = targetHost.to_local(global_pos)
-			# The target position is always local coordinates 0,0 of the final container
-			_target_position = targetHost.get_stack_position(self)
-			state = MOVING_TO_CONTAINER
-			if not targetHost.faceup_cards:
-				if set_is_faceup(false) == _ReturnCode.FAILED:
+			# The below checks if the container we're moving is in popup
+			# If the card is also in a popup, we assume we're moving back
+			# to the same container, so we do nothing
+			# The finite state machine  will reset the card to its position
+			if "CardPopUpSlot" in parentHost.name:
+				if targetHost.get_node("ViewPopup").visible == true:
+					pass
+			else:
+				# Added because sometimes it ended up stuck and a card remained
+				# visible on top of deck
+				$Tween.remove_all()
+				# We need to adjust the end position based on the local rect inside
+				# the container control node
+				# So we transform global coordinates to container rect coordinates.
+				previous_pos = targetHost.to_local(global_pos)
+				# The target position is always local coordinates 0,0
+				# of the final container
+				_target_position = targetHost.get_stack_position(self)
+				state = MOVING_TO_CONTAINER
+				if set_is_faceup(targetHost.faceup_cards) == _ReturnCode.FAILED:
 					print("ERROR: Something went unexpectedly in set_is_faceup")
-			# If we have fancy movement, we need to wait for 2 tweens to finish
-			# before we reorganize the stack.
-			# One for the fancy move, and then the move to the final position
-			yield($Tween, "tween_all_completed")
-			if cfc.fancy_movement:
+				# If we have fancy movement, we need to wait for 2 tweens to finish
+				# before we reorganize the stack.
+				# One for the fancy move, and then the move to the final position
 				yield($Tween, "tween_all_completed")
-			targetHost.reorganize_stack()
+				if cfc.fancy_movement:
+					yield($Tween, "tween_all_completed")
+				targetHost.reorganize_stack()
 		else:
 			interruptTweening()
 			if len(_potential_cards):
@@ -1141,8 +1149,10 @@ func _determine_idle_state() -> void:
 		state = IN_HAND
 	# The extra if is in case the ViewPopup is currently active when the card
 	# is being moved into the container
-	elif get_parent() in cfc.piles or "CardPopUpSlot" in get_parent().name:
+	elif get_parent() in cfc.piles:
 		state = IN_PILE
+	elif "CardPopUpSlot" in get_parent().name:
+		state = IN_POPUP
 	else:
 		state = ON_PLAY_BOARD
 
@@ -1636,7 +1646,7 @@ func _process_card_state() -> void:
 							position, intermediate_position, 0.25,
 							Tween.TRANS_CUBIC, Tween.EASE_OUT)
 					yield($Tween, "tween_all_completed")
-					if not scale.is_equal_approx(cfc.PLAY_AREA_SCALE):
+					if not scale.is_equal_approx(Vector2(1,1)):
 						$Tween.remove(self,'scale')
 						$Tween.interpolate_property(self,'scale',
 								scale, Vector2(1,1), 0.5,
@@ -1656,6 +1666,19 @@ func _process_card_state() -> void:
 		IN_PILE:
 			set_focus(false)
 			set_mouse_filters(false)
+			if scale != Vector2(1,1):
+				scale = Vector2(1,1)
+		IN_POPUP:
+			# We make sure that a card in a popup stays in its position
+			# Unless moved
+			set_focus(false)
+			set_mouse_filters(true)
+			if modulate[3] != 1:
+				modulate[3] = 1
+			if scale != Vector2(0.75,0.75):
+				scale = Vector2(0.75,0.75)
+			if position != Vector2(0,0):
+				position = Vector2(0,0)
 		FOCUSED_IN_POPUP:
 			# Used when the card is displayed in the popup grid container
 			set_focus(true)
